+++ 2COMMONPROJECTS/audio-services-test/src/test/uk/co/modularaudio/service/rendering/config/RenderingTestConfig.java
@@ -80,7 +80,7 @@ public class RenderingTestConfig
 	}
 
 	@Override
-	public long getCurrentUiFrameTime()
+	public int getCurrentUiFrameTime()
 	{
 		return 0;
 	}
